linode-cli: Remove update check
Upon every API call, this program will check the API version and make a
request to PyPi to check the latest version and, if not latest, will
show a warning like

> The API responded with version 1.2.4, which is newer than the CLI's
> version of 1.2.3.  Please update the CLI to get access to the newest
> features.  You can update with a simple `pip install --upgrade
> linode-cli`

This is not really useful with Nix, as the update message is wrong. And
connecting to a third-party on each usage is not desired, either.