fcitx5-rime: add an environment variable for shared data dir passing librime.
The original RIME_DATA_DIR is fixed at build time which is set to the package
dir under nix store. A such dir is read only for users so if some one want to
add a shared data, he has to override fcitx5-rime or librime, which isn't
user-friendly.

On other distributions, RIME_DATA_DIR may be under /usr/local/share/fcitx5 which
is not existed on NixOS. If we set its to /run/current-system/sw/share/..., this
will break for non-NixOS user.

Therefore, we patch it with a special environment variable to let user to load
it dynamically.

Co-Authored-by: Lan Tian <xuyh0120@outlook.com>