gnome.eog: Hardcode library path in gir
By default, gobject-introspection refers to libraries using their
file name in the GIR file. Since we do not have FHS to reliably
find libraries by their name, we are patching gobject-introspection
to use absolute paths.

This works fine for libraries installed to "$out/lib" but when a different
path is used, the hard-coded absolute path will be incorrect.
To deal with this case, we also patch Meson to be able
to pass the install_dir to gobject-introspection.
But it currently only passes the data for targets
that have absolute install_dir set.

To have correct libeog path, we therefore need to absolutize the install_dir.