libtool: fix /usr/bin/file impurity
libtool's libtool.m4 script assumes that `file` is available, and can
be found at `/usr/bin/file` (this path is hardwired).  Furthermore,
the script with this assumption is vendored into the ./configure
scripts of an enormous number of packages.

The use of /usr/bin/file in ./configure scripts only affects sandboxed
cross-builds to a x86_64-linux host as well as any sandboxed build
(cross or native) to these hosts: x86_64-freebsd, *-hpux, *-irix,
mips64*-linux, powerpc*-linux, s390x-linux, s390x-tpf, sparc-linux,
and *-solaris.

This commit adds `replaceUsrBinFile`, which will check the ./configure
script to see if it vendors a pre-version-2.4.7 copy of libtool.m4.
If it does, it will `substituteInPlace` the use of `/usr/bin/file`
with `${buildPackages.file}/bin/file`.

If a pre-version-2.4.7 copy of libtool.m4 is *not* vendored then
replaceUsrBinFile will cause the build to fail -- this ensures that
replaceUsrBinFile gets removed as upstream packages gradually upgrade
to a fixed version of libtool and re-run autoconf.

This commit allows the following commands to complete, which should
enable Hydra to produce bootstrap-files for mips64el:

```
  nix-build \
    --option sandbox true \
    --option sandbox-fallback false \
    pkgs/top-level/release-cross.nix \
    -A bootstrapTools.mips64el-linux-gnuabi64.build

  nix-build \
    --option sandbox true \
    --option sandbox-fallback false \
    . \
    -A pkgsCross.mips64el-linux-gnuabi64.nix_2_4
```

Co-authored-by: Sandro <sandro.jaeckel@gmail.com>