qt5*.qtwayland: deconflict patching
This fixes build in qt514 case.

The usual way here is to provide patches for each qt5 version
separately.  No other module adds them in this generic way.
The problem is when you combine the approaches; qtModule will only
take the list from the module and ignore the version-specific list.