pkgs/errcheck: 1.6.0 -> 20220325-e62617a, and go1.18
This updates errcheck to a newer version which has go1.18 support.

There's no tagged release yet for this, but without this change, this
package breaks under go1.18, so it seems worth updating before upstream
tags.

Related to #164320

The switch to 'subPackages' was added because 'testdata' contains a
'main_test.go' file which is not meant to be actually tested, and the
default checkPhase will try to run 'cd testdata && go test' unless you
do _something_ to override it. Since checkPhase doesn't respect
'excludedPackages', but does respect 'subPackages', this seemed like a
reasonable way to do that.