talosctl: provide path to UEFI firmware
By default, a `talosctl cluster create` will spin up VMs with UEFI.
For this to work, an UEFI firmware binary needs to be passed to qemu.
`talosctl` looks for these in various places in `/usr/share`.

On NixOS, these paths don't exist. While it's possible to pass
additional search paths (via the `--extra-uefi-search-paths` cmdline
arg), it's much more convenient to automatically have a Nix-built OVMF
firmware in the search paths.

Right now, we only do this for the "native" OVMF binaries - so if you
run an x86_64-linux talosctl binary, you can spin up a x86_64 VM out of
the box, but you won't be able to start a aarch64-linux VM without
explicitly setting `--extra-uefi-search-paths` to a aarch64 firmware
path - this is mostly due to the OVMF package not properly
crosscompiling.

We only add to the default search paths, not replace them, so people
running talosctl on non-NixOS have more search paths available.