python3Packages.fipy: fix broken build
See #167742

python3Packages.fipy is currently broken on Hydra,
https://hydra.nixos.org/build/172501646. FiPy is unable to read the
version string for Gmsh. This is currently fixed on FiPy's master and
so appropriate patch is now applied.. Patch should be removed at next
release.

Also, Gmsh was not in the path when running the tests so now added to
checkInputs.