ipxe: 1.21.1 -> unstable-2022-04-06
iPXE does not make stable releases[1].  The tag we were using was an
anomoly[2], and is from 2020.

The geniso script has been replaced with a generic genfsimg script
that works differently.  There are no longer variables we can set to
tell it where to find isolinux.bin or ldlinux.c32, so we have to patch
the script.  It'll look in the same place for both files.

The --preserve=mode modification doesn't seem to be necessary any more
— when it was introduced in 4d7bea7e5ef ("ipxe: fix ISO build"), it
was required to make a working ISO with an embedded script.  But my
ISO with an embedded script now works fine without that change.

[1]: https://github.com/ipxe/ipxe/discussions/544
[2]: https://github.com/ipxe/ipxe/discussions/544#discussioncomment-1894506