openjdk: add support for powerpc64le
This commit adds entries to the Rosetta Stones in
adoptopenjdk-bin/generate-sources.py and compilers/openjdk/8.nix, and
runs adoptopenjdk-bin/generate-sources.py to regenerate
adoptopenjdk-bin/sources.json.

With this commit, `nix-build . -A jdk8_headless` succeeds on
powerpc64le.  Headless jdk is used as part of the build process for
many packages so this opens up access to them.