CODEOWNERS: Add layus for autoPatchelfHook
With the re-implementation in Python merged[1], it no longer makes sense
for me to track issues and pull requests. I did this originally because
people were forgetting (rightfully so) to run tests against all that
proprietary stuff we have in nixpkgs that is using autoPatchelfHook.

We still can't test these automatically but with me no longer being the
author of the code, I hereby drop my entry in CODEOWNERS and instead
replace it with layus, who's the author of the rewrite.

[1]: https://github.com/NixOS/nixpkgs/pull/149731

Signed-off-by: aszlig <aszlig@nix.build>