snes9x: init at 1.61
This is done by modifying the current `snes9x-gtk` derivation, including
a new `withGtk` option that when enabled will build the `snes9x-gtk`,
but when disabled will build `snes9x` instead. `snes9x` has less
dependencies than `snes9x-gtk`, having only X11 as a major dependency
and a minor one in ALSA. However, the only audio backend supported is ALSA,
so audio is only available on Linux.

Still, this has less dependencies than the GTK version, so it should be
useful as a minimal version for cross-compilation cases or when
porting to new devices. This also builds in macOS, however as I said
before, shouldn't have any sound there.