lshw: rename version to fix repology listing
`lshw` is not showing in the nixpkgs-unstable listings in repology, but
it is showing in the previous stable release (release-21.11). The reason
for this was a change in 8d0267dc8f02e010785d6f90573ea1135b7957e7, that
fixed the `lshw` version to include its letter prefix.

However, the way the version is computed for repology is to parse the
`name` attr instead, separating it in `pname` and `version`. The
function that does this (`builtins.parseDrvName`) considers anything
that is a `name` everything up to the first dash followed by a digit.
Because the `version` includes the letter `B` as prefix, it them ends up
splitting it wrong.

See https://github.com/NixOS/nix/pull/4463 for a proper fix, but for now
let's just "fix" this by not including the prefix in the `version`.