python3Packages.dropbox: Update hash for re-release of 11.30.0
The version of the package was not set properly and defaulted to 0.0.0 .
The package was re-released by upstream (with the same version number)
after it was reported.

This package will now have a correct version number in the dist-info when built,
fixing packages that require it to be within certain version bounds.