godot-export-templates: strip export template file
Stripping reduces the template size from around 500MB to 40MB for Linux.
This is added explicitly here because mkDerivation does not automatically
strip binaries in the template directory.

This also impacts the size of the exported games and programs.
For example, here are the size of the `oh-my-git` package using this
template, without and with stripping:

```
❯ du -sh result*/
375M	result-oh-my-git-without-strip/
45M	result-oh-my-git-with-strip/
```

GitHub: closes https://github.com/NixOS/nixpkgs/issues/170470