dsq: replace package test with upstream test script
The package has its own custom test data and suite. There are no Go
tests so run the custom suite in `checkPhase` instead.

The test script wants to download a big CSV file from an S3 bucket,
which won’t work in a sandbox. To keep that working We’ll use instead
use `fetchurl` and point curl to the downloaded file in the Nix store.