zq: init at v1.2.0
This change adds zq, a part of the zed project https://zed.brimdata.io/.

zq is a command-line tool that uses the Zed language for pipeline-style
search and analytics. zq can query a variety of data formats in files,
over HTTP, or in S3 storage. It is particularly fast when operating on
data in the Zed-native ZNG format.

The zq design philosophy blends the query/search-tool approach of jq,
awk, and grep with the command-line, embedded database approach of
sqlite and duckdb.