.github/stale bot: stop commenting
Niklas Hambüchen noted in

https://github.com/NixOS/rfcs/pull/124#issuecomment-1113647856

that we can add a label description for the stale label that appears
when you hover over the label, or look on

https://github.com/NixOS/nixpkgs/labels

I set the description to be

> https://github.com/NixOS/nixpkgs/blob/master/.github/STALE-BOT.md

which is our page explaining the stale bot.

The stale bot comments/emails are a significant burden on our most
prodigious contributors, and the reason for their existence to orient
new contributors.

Since our stale bot's configuration is benign enough to ignore (it
does not close), I believe it is good enough to satisfy the new
contributor orientation with the label description.

Therefore, this commit disables commenting when labeling an issue or
PR stale.