rabbitmq-server: Don't look for plugins in the store path
By default RabbitMQ searches for plugins in its store path. This can
only work for plugins which are built as part of the server, because
the store is immutable and other packages are not allowed to install
files there.

This change sets the plugin search path to point into the overlying
symlinkJoined layer, so that out-of-tree plugins can be built,
packaged and installed independently.