libxcrypt: prevent build failure on warnings
The upstream autoconf build system has a default setting that treats
compiler warnings as errors. While this can be useful for development
and maintenance it's not ideal for packaging and distribution: newer
versions of compilers may detect warnings that previous versions of the
compilers failed to detect and cause this package to fail to build.